+++ src/components/experience/Experience.js
@@ -5,7 +5,7 @@ const Experience = () => {
   return (
     <div className="experience">
       <div className="d-flex justify-content-center my-5">
-        <h1>Experience</h1>
+        <h1 className="experience-heading">Experience</h1>
       </div>
       <div className="container experience-wrapper">
         <div className="timeline-block timeline-block-right">
