+++ src/components/portfolio/Portfolio.js
@@ -1,145 +1,102 @@
 import React from "react";
-import netflix from "../images/netflix.png";
-import cityGuide from "../images/city-guide-app.png";
-import portfolio from "../images/portfolio.png";
-import taskManager from "../images/task-manager.png";
+import banyan from "../../img/banyanlabs.png";
+import diver from "../../img//scubaSite.png";
+
 // FONTAWESOME IMPORTS
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSearchPlus } from "@fortawesome/free-solid-svg-icons";
 // REACT POPUPBOX
 import { PopupboxManager, PopupboxContainer } from "react-popupbox";
 import "react-popupbox/dist/react-popupbox.css";
-
+import './portfolio.css';
 
 
 const Pofrfolio = () => {
 
-  // Netflix
-  const openPopupboxNetflix = () => {
+  // Banyan Labs
+  const openPopupboxBanyan = () => {
     const content = (
       <div>
-        <img className="portfolio-image-popupbox" src={netflix} alt="Netflix Clone Project..." />
-        <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Ex cumque illo est expedita quos adipisci suscipit unde itaque qui perferendis.</p>
-        <b>GitHub:</b> <a className="hyper-link" onClick={() => window.open("https://github.com/8020Coding/netflix-project")}>https://github.com/8020Coding/netflix-project</a>
+        <img className="portfolio-image-popupbox" src={banyan} alt="Banmyan Web Site..." />
+        <p>My first project, after getting hired on at Banyan Labs, was to help the web teasm build the Banyan webste using React.</p>
+        <b>Web:</b> <a className="hyper-link" onClick={() => window.open("https:/banyanlabs.io")}>https://banyanlabs.io</a>
       </div>
     )
     PopupboxManager.open({ content })
   }
 
-  const popupboxConfigNetflix = {
+  const popupboxConfigBanyan = {
     titleBar: {
       enable: true,
-      text: "Netflix clone project."
+      text: "Banyan Labs Website."
     },
     fadeIn: true,
     fadeInSpeed: 500
   }
 
-  // City Guide App
-  const openPopupboxCityGuide = () => {
+//
+  const openPopupboxScubaBlog = () => {
     const content = (
       <div>
-        <img className="portfolio-image-popupbox" src={cityGuide} alt="City Guide App Project..." />
-        <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Ex cumque illo est expedita quos adipisci suscipit unde itaque qui perferendis.</p>
-        <b>Demo:</b> <a className="hyper-link" onClick={() => window.open("https://city-guide-app-project.herokuapp.com/", "_blank")}>https://city-guide-app-project.herokuapp.com/</a>
-        <br />
-        <b>GitHub:</b> <a className="hyper-link" onClick={() => window.open("https://github.com/8020Coding/city-guide-app", "_blank")}>https://github.com/8020Coding/city-guide-app</a>
-      </>
-    )
-    PopupboxManager.open({ content })
-  }
-
-  const popupboxConfigCityGuide = {
-    titleBar: {
-      enable: true,
-      text: "City Guide App project."
-    },
-    fadeIn: true,
-    fadeInSpeed: 500
-  }
-
-  // Portfolio Project
-  const openPopupboxPortfolio = () => {
-    const content = (
-      <div>
-        <img className="portfolio-image-popupbox" src={portfolio} alt="Portfolio Project..." />
-        <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Ex cumque illo est expedita quos adipisci suscipit unde itaque qui perferendis.</p>
-        <b>Demo:</b> <a className="hyper-link" onClick={() => window.open("https://portfolio-rea-and-material-ui.herokuapp.com/", "_blank")}>https://portfolio-rea-and-material-ui.herokuapp.com/</a>
-        <br />
-        <b>GitHub:</b> <a className="hyper-link" onClick={() => window.open("https://github.com/8020Coding/portfilio", "_blank")}>https://github.com/8020Coding/portfilio</a>
+        <img className="portfolio-image-popupbox" src={diver} alt="Django Scuba Blog project." />
+        <p>Scuba diving website with login and registration verification using Django. It also has a blog section where scuba divers can blog about there best dive spots or deepest dives, etc.. Has profile section and if you click on person name it shows all of that persons blogs from newest to oldest.</p>
+        
+        <b>GitHub:</b> <a className="hyper-link" onClick={() => window.open("https://github.com/joseph-caldwell/Deep-Divers/tree/main/scuba_site", "_blank")}>https://github.com/joseph-caldwell/Deep-Divers/tree/main/scuba_site</a>
       </div>
     )
     PopupboxManager.open({ content })
   }
 
-  const popupboxConfigPortfolio = {
+  const popupboxConfigScubaBlog = {
     titleBar: {
       enable: true,
-      text: "Poftfolio React and Material UI project."
+      text: "Django Scuba Blog project."
     },
     fadeIn: true,
     fadeInSpeed: 500
   }
 
-  // Task Manager React and Redux Project
-  const openPopupboxTaskManager = () => {
-    const content = (
-      <div>
-        <img className="portfolio-image-popupbox" src={taskManager} alt="Task Manager React and Redux Project..." />
-        <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Ex cumque illo est expedita quos adipisci suscipit unde itaque qui perferendis.</p>
-        <b>Demo:</b> <a className="hyper-link" onClick={() => window.open("https://react-redux-task-manager.herokuapp.com/", "_blank")}>https://react-redux-task-manager.herokuapp.com/</a>
-        <br />
-        <b>GitHub:</b> <a className="hyper-link" onClick={() => window.open("https://github.com/8020Coding/task-manager", "_blank")}>https://github.com/8020Coding/task-manager</a>
-      </div>
-    )
-    PopupboxManager.open({ content })
-  }
 
-  const popupboxConfigTaskManager = {
-    titleBar: {
-      enable: true,
-      text: "Task Manager React and Redux project."
-    },
-    fadeIn: true,
-    fadeInSpeed: 500
-  }
+
 
 
   return (
     <div className="portfolio-wrapper">
-      <div className="container">
-        <h1 className="text-uppercase text-center py-5">portfolio</h1>
-        <div className="image-box-wrapper row justify-content-center">
-          <div className="portfolio-image-box" onClick={openPopupboxNetflix}>
-            <img className="portfolio-image" src={netflix} alt="Netflix Clone Project..." />
-            <div className="overflow"></div>
-            <FontAwesomeIcon className="portfolio-icon" icon={faSearchPlus} />
+        <div className="container">
+          <h1 className="portfolio-heading text-center py-5">portfolio</h1>
+            <div className="image-box-wrapper row justify-content-center">
+              <div className="portfolio-image-box" onClick={openPopupboxBanyan}>
+                <img className="portfolio-image" src={banyan} alt="Banyan Labs Web Site..." />
+            //   <div className="overflow"></div>
+            // <FontAwesomeIcon className="portfolio-icon" icon={faSearchPlus} />
           </div>
           {/* - */}
-          <div className="portfolio-image-box" onClick={openPopupboxCityGuide}>
-            <img className="portfolio-image" src={cityGuide} alt="City Guide Project..." />
-            <div className="overflow"></div>
-            <FontAwesomeIcon className="portfolio-icon" icon={faSearchPlus} />
+          <div className="portfolio-image-box" onClick={openPopupboxScubaBlog}>
+            <img className="portfolio-image" src={diver} alt="City Guide Project..." />
+            // <div className="overflow"></div>
+            // <FontAwesomeIcon className="portfolio-icon" icon={faSearchPlus} />
           </div>
           {/* - */}
-          <div className="portfolio-image-box" onClick={openPopupboxPortfolio}>
-            <img className="portfolio-image" src={portfolio} alt="Portfolio React and Material UI Project..." />
-            <div className="overflow"></div>
-            <FontAwesomeIcon className="portfolio-icon" icon={faSearchPlus} />
+          
+          <div className="portfolio-image-box" onClick={openPopupboxBanyan}>
+            <img className="portfolio-image" src={banyan} alt="Banyan Labs Web Site..." />
+            // <div className="overflow"></div>
+            // <FontAwesomeIcon className="portfolio-icon" icon={faSearchPlus} />
           </div>
           {/* - */}
-          <div className="portfolio-image-box" onClick={openPopupboxTaskManager}>
-            <img className="portfolio-image" src={taskManager} alt="Task Manager React and Redux Project..." />
-            <div className="overflow"></div>
-            <FontAwesomeIcon className="portfolio-icon" icon={faSearchPlus} />
+          <div className="portfolio-image-box" onClick={openPopupboxScubaBlog}>
+            <img className="portfolio-image" src={diver} alt="City Guide Project..." />
+            // <div className="overflow"></div>
+            // <FontAwesomeIcon className="portfolio-icon" icon={faSearchPlus} />
           </div>
-        </div>
+   
       </div>
-      <PopupboxContainer {...popupboxConfigNetflix} />
-      <PopupboxContainer {...popupboxConfigCityGuide} />
-      <PopupboxContainer {...popupboxConfigPortfolio} />
-      <PopupboxContainer {...popupboxConfigTaskManager} />
+      <PopupboxContainer {...popupboxConfigBanyan} />
+      <PopupboxContainer {...popupboxConfigScubaBlog} />
+      <PopupboxContainer {...popupboxConfigBanyan} />
+      <PopupboxContainer {...popupboxConfigScubaBlog} />
     </div>
+     </div>
   )
 }
