+++ src/components/portfolio/portfolio.css
@@ -1,92 +1,120 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto+Condensed:wght@300&display=swap');
+
+:root {
+  --primary-white: #ffffff;
+  --primary-light-grey: #f4f4f4;
+  --primary-grey: #ccd1d9;
+  --primary-dark-grey: #666666;
+  --primary-orange: #f9ab00;
+  --primary-dark: rgb(52, 58, 64);
+  --primary-red: #ef4035;
+  --primary-hover-red: #f33501;
+  --primary-black: #000000;
+  --primary-transparent-black: rgba(0,0,0,0.7);
+  --primary-green: green;
+}
+
+body {
+font-family: "Roboto Condensed", sans-serif !important;
+}
+
+
 /************************ PORTFOLIO *********************************/
 .portfolio-wrapper {
-    background: var(--primary-light-grey);
-    padding: 3rem 0;
-  }
-  
-  .portfolio-wrapper h1 {
-    color: var(--primary-orange);
-  }
-  
-  .portfolio-image {
-    width: 15rem;
-    border: 1px solid var(--primary-orange);
-    padding: 0 0.5rem;
-  }
-  
-  .portfolio-image-box {
-    position: relative;
-    margin: 1rem;
-  }
-  
-  /*------------ OVERFLOW BOX ------------------*/
-  .overflow {
-    position: absolute;
-     opacity: 0;
-     background: var(--primary-black);
-     width: 15rem;
-     height: 8rem;
-     top: 0;
-    border: 1px solid var(--primary-red);
-    cursor: pointer;
-  }
-  
-  .portfolio-image-box:hover .overflow {
-    opacity: 60%;
-    transition: 0.3s ease-in-out;
-  }
-  
-  .portfolio-icon {
-    position: absolute;
-    opacity: 0;
-    margin-left: -8rem;
-    margin-top: 3.6rem;
-    color: var(--primary-orange);
-    font-size: 1.5rem;
-  }
-  
-  .portfolio-image-box:hover .portfolio-icon {
-    opacity: 70%;
-    transition: 0.3s ease-in-out;
-    cursor: pointer;
-  }
-  
-  /*------------ OVERFLOW BOX ------------------*/
+  background: var(--primary-light-grey);
+  padding: 3rem 0;
+}
+
+.portfolio-wrapper h1 {
+  color: var(--primary-orange);
+}
+
+.portfolio-heading{
+  color: var(--primary-orange);
+  text-transform: uppercase;
+  font-family:algerian !important;
+  text-decoration: underline;
+}
+
+.portfolio-image {
+  width: 15rem;
+  border: 1px solid var(--primary-orange);
+  padding: 0 0.5rem;
+}
+
+.portfolio-image-box {
+  position: relative;
+  margin: 1rem;
+}
+
+/*------------ OVERFLOW BOX ------------------*/
+.overflow {
+  position: absolute;
+   opacity: 0;
+   background: var(--primary-black);
+   width: 15rem;
+   height: 8rem;
+   top: 0;
+  border: 1px solid var(--primary-red);
+  cursor: pointer;
+}
+
+.portfolio-image-box:hover .overflow {
+  opacity: 60%;
+  transition: 0.3s ease-in-out;
+}
+
+.portfolio-icon {
+  position: absolute;
+  opacity: 0;
+  margin-left: -8rem;
+  margin-top: 3.6rem;
+  color: var(--primary-orange);
+  font-size: 1.5rem;
+}
+
+.portfolio-image-box:hover .portfolio-icon {
+  opacity: 70%;
+  transition: 0.3s ease-in-out;
+  cursor: pointer;
+}
+
+/*------------ OVERFLOW BOX ------------------*/
+.portfolio-image-popupbox {
+  width: 45rem;
+  padding: 0 0.5rem;
+}
+
+.hyper-link {
+  cursor: pointer; 
+  color: var(--primary-orange);
+}
+
+.hyper-link:hover {
+  color: var(--primary-red);
+}
+
+/*------------ MEDIA ------------------*/
+@media(max-width: 768px) {
   .portfolio-image-popupbox {
-    width: 45rem;
-    padding: 0 0.5rem;
+    width: 100%;
   }
-  
-  .hyper-link {
-    cursor: pointer; 
-    color: var(--primary-orange);
+}
+
+@media(max-height: 570px) {
+.popupbox-wrapper {
+  height: 100%;
+}
+  .portfolio-image-popupbox {
+    width: 50%;
   }
-  
-  .hyper-link:hover {
-    color: var(--primary-red);
-  }
-  
-  /*------------ MEDIA ------------------*/
-  @media(max-width: 768px) {
-    .portfolio-image-popupbox {
-      width: 100%;
-    }
-  }
-  
-  @media(max-height: 570px) {
+}
+
+@media(max-height: 280px) {
   .popupbox-wrapper {
-    height: 100%;
+    height: 90%;
   }
     .portfolio-image-popupbox {
-      width: 50%;
+      width: 30%;
     }
   }
-  
-  @media(max-height: 280px) {
-    .popupbox-wrapper {
-      height: 90%;
-    }
-      .portfolio-image-popupbox {
-        width: 30%;
-      }
-    }
